// src/KRINT.Frontend/src/app/api/model/dashboardStatsDto.ts
/**
 * KRINT.API | v1
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { DashboardStatsDtoTotalCpuPercent } from './dashboardStatsDtoTotalCpuPercent';
import { ActivityEntryDto } from './activityEntryDto';
import { DashboardStatsDtoTotalInstances } from './dashboardStatsDtoTotalInstances';
import { EngineCountDto } from './engineCountDto';
import { BackupEntryDtoSizeBytes } from './backupEntryDtoSizeBytes';


export interface DashboardStatsDto { 
    totalInstances: DashboardStatsDtoTotalInstances;
    runningInstances: DashboardStatsDtoTotalInstances;
    totalMemoryBytes: BackupEntryDtoSizeBytes;
    totalCpuPercent: DashboardStatsDtoTotalCpuPercent;
    perEngine: Array<EngineCountDto>;
    recentActivity: Array<ActivityEntryDto>;
}